Refactor delta-mutators to work from an initial trn delta.
This allows deltas to be grouped efficiently - continuing from an existing
delta rather than creating a new delta to then be converged with the original.

The cap is trn, so that the delta itself may become sendable by consuming to val,
if the goal is to transmit the delta in Pony instead of serializing it to bytes
to be transmitted over a network. With trn, either is possible with little hassle.